+++ app/src/main/java/cn/spacexc/wearbili/dataclass/watchlater/Owner.kt
@@ -2,6 +2,6 @@ package cn.spacexc.wearbili.dataclass.watchlater
 
 data class Owner(
     val face: String,
-    val mid: Int,
+    val mid: Long,
     val name: String
 )
